+++ lib/archethic.dart
@@ -1,0 +1,11 @@
+/// Package archEthic aims to provide a easy way to create ArchEthic transaction and to send them over the network.
+///
+/// This implementation is based on Official ArchEthic Javascript library for Node and Browser.
+library archethic;
+
+export 'src/crypto.dart';
+export 'src/transaction_builder.dart';
+export 'src/utils.dart';
+export 'src/services/address_service.dart';
+export 'src/services/api_coins_service.dart';
+export 'src/services/api_service.dart';