+++ test/crypto_test.dart
@@ -9,9 +9,9 @@ import 'dart:typed_data';
 import 'package:test/test.dart';
 
 // Project imports:
-import 'package:archethic_lib_dart/crypto.dart' as crypto;
-import 'package:archethic_lib_dart/model/key_pair.dart';
-import 'package:archethic_lib_dart/utils.dart';
+import 'package:archethic_lib_dart/src/crypto.dart' as crypto;
+import 'package:archethic_lib_dart/src/model/key_pair.dart';
+import 'package:archethic_lib_dart/src/utils.dart';
 
 void main() {
   group('Crypto', () {